Re-organized cache folders
