+++ packages/suspense/src/cache/createRangeCache/createRangeCache.ts
@@ -1,18 +1,22 @@
-import { STATUS_PENDING, STATUS_REJECTED, STATUS_RESOLVED } from "../constants";
+import {
+  STATUS_PENDING,
+  STATUS_REJECTED,
+  STATUS_RESOLVED,
+} from "../../constants";
 import {
   ComparisonFunction,
   PendingRecord,
   RangeCache,
   Record,
   Thenable,
-} from "../types";
-import { assertPendingRecord } from "../utils/assertPendingRecord";
-import { createDeferred } from "../utils/createDeferred";
-import { defaultGetKey } from "../utils/defaultGetKey";
-import { findMissingRanges } from "../utils/findMissingRanges";
-import { findNearestIndexAfter } from "../utils/findNearestIndex";
-import { isPendingRecord } from "../utils/isPendingRecord";
-import { sliceValues } from "../utils/sliceValues";
+} from "../../types";
+import { assertPendingRecord } from "../../utils/assertPendingRecord";
+import { createDeferred } from "../../utils/createDeferred";
+import { defaultGetKey } from "../../utils/defaultGetKey";
+import { isPendingRecord } from "../../utils/isPendingRecord";
+import { findMissingRanges } from "./findMissingRanges";
+import { findNearestIndexAfter } from "./findNearestIndex";
+import { sliceValues } from "./sliceValues";
 
 // Enable to help with debugging in dev
 const DEBUG_LOG_IN_DEV = false;
