+++ packages/suspense/src/cache/createRangeCache/findNearestIndex.ts
@@ -1,5 +1,5 @@
-import { GetPoint } from "../cache/createRangeCache";
-import { ComparisonFunction } from "../types";
+import { GetPoint } from "./createRangeCache";
+import { ComparisonFunction } from "../../types";
 
 export function findNearestIndex<Point, Value>(
   sortedValues: Value[],
